requirements.txt
# Core AI / ML
openai-whisper==20231117
openai>=1.3.0

# Audio / Video processing
ffmpeg-python==0.2.0
pydub==0.25.1

# Web UI
streamlit>=1.28.0

# Document export
fpdf2==2.7.6
python-docx==1.1.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
numpy>=1.24.0

# Optional: speaker diarization (requires torch)
# pyannote.audio==3.1.1
# torch>=2.0.0
